Added option to set K8sGraphFlags for tests
