+++ checkov/kubernetes/graph_manager.py
@@ -5,7 +5,8 @@ from typing import Any, TYPE_CHECKING
 from checkov.common.graph.db_connectors.db_connector import DBConnector
 from checkov.common.graph.graph_manager import GraphManager
 from checkov.kubernetes.graph_builder.local_graph import KubernetesLocalGraph
-from checkov.kubernetes.kubernetes_utils import get_folder_definitions, K8sGraphFlags
+from checkov.kubernetes.kubernetes_utils import get_folder_definitions
+from checkov.kubernetes.kubernetes_graph_flags import K8sGraphFlags
 
 if TYPE_CHECKING:
     from networkx import DiGraph
