+++ checkov/kubernetes/kubernetes_utils.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import logging
 import os
 from copy import deepcopy
-from dataclasses import dataclass
 from typing import Dict, Any
 
 import dpath
@@ -229,15 +228,3 @@ def remove_metadata_from_attribute(attribute: dict[str, Any] | None) -> None:
     if isinstance(attribute, dict):
         attribute.pop("__startline__", None)
         attribute.pop("__endline__", None)
-
-
-@dataclass()
-class K8sGraphFlags:
-    create_complex_vertices: bool
-    create_edges: bool
-
-    def __init__(self, create_complex_vertices: bool = False, create_edges: bool = False) -> None:
-        create_complex_vertices_env_var: bool = bool(os.environ.get('CREATE_COMPLEX_VERTICES'))
-        create_edges_env_var: bool = bool(os.environ.get('CREATE_EDGES'))
-        self.create_complex_vertices = create_complex_vertices or create_complex_vertices_env_var
-        self.create_edges = create_edges or create_edges_env_var
